[common] Improve SystemD's own logging by using "combined" format
This implements https://fedoraproject.org/wiki/Changes/Unit_Names_in_Systemd_Messages
on all RPS hosts and containers, for improved readability/debuggability
of logs when using templated units - it will log the actual unit names.

This configuration option is available starting from systemd version 249.
As of 2022-05, this means it will only work on Debian systems where at
least the bullseye-backports repository enabled.
(This commit does not enforce that - this would need to be decided
on a per-role basis or globally.)

On very old systems (e.g. Debian buster without backports) this will
throw a trivial warning, as the whole "StatusUnitFormat" option is
unknown. This can safely be ignored.