remove image name
apparently ansible docker compose logic does not work properly with build
and pull at the same time. if you specify an image name it will try to pull
the image (which might not yet be there).

so remove the image name (which is not really necessary in the build case
anyway)